Nest graphql endpoint under api
