+++ app/controllers/api/api_controller.rb
@@ -1,0 +1,4 @@
+module Api
+  class ApiController < ActionController::API
+  end
+end
