+++ config/routes.rb
@@ -1,4 +1,5 @@
 Rails.application.routes.draw do
-  post "/graphql", to: "graphql#execute"
-  # For details on the DSL available within this file, see https://guides.rubyonrails.org/routing.html
+  namespace :api do
+    post "/graphql", to: "graphql#execute"
+  end
 end
